feat(tui): currently functional but messy TUI code
This is a working prototype for the TUI, it needs a lot of
refactoring work done to be presentable for a release but as a proof of
concept but it functions.

Adding a feed and storing any type of feed data is not implemented,
refreshing is not currently available, and there is no content
formatting so a lot of `&lt;p&gt:`.