fix(greyxml): update text spans to correctly hold start row and column
This means that the span for a multiline text is holding the start of
the row and column rather than incorrectly going to the end and then
trying to browse backwards.

This closes #7 without needing the compicated logic mentioned.

This could use a refactor, there is a lot of duplication for every
Lexicals::Text.